third_party/WebKit/LayoutTests/platform/win/virtual/threaded/fast/scroll-behavior/scroll-customization/scrollstate-distribute-to-scroll-chain-descendant-expected.txt

CONSOLE ERROR: line 30: Uncaught TypeError: element.setDistributeScroll is not a function
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: element.setDistributeScroll is not a function
FAIL These tests only work with scroll customization enabled. assert_true: 'ScrollState' in window expected true got false
Harness: the test ran to completion.
